requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect2==0.5.0
dateparser==1.1.0
defusedxml==0.7.1
Django
django-ckeditor
django-cors-headers==3.11.0
django-filter==21.1
django-js-asset==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
easy-thumbnails==2.8.1
frozenlist==1.3.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.0.1
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-binance==1.0.15
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
websockets==9.1
yarl==1.7.2
